+++ Game/Point.h
@@ -4,7 +4,7 @@ using namespace std;
 
 class Point {
 public:
-	int x, y;
-	Point(int xPosition, int yPosition);
+	double x, y;
+	Point(double xPosition, double yPosition);
 	Point() : Point(0, 0) {};
 };
